Extract Endpoint Instance List component
