+++ src/ServicePulse.Host/vue/src/components/monitoring/EndpointInstances.vue
@@ -1,0 +1,192 @@
+<script setup>
+import { ref, onMounted } from "vue";
+import { useRouter } from "vue-router";
+import { formatGraphDecimal, formatGraphDuration, smallGraphsMinimumYAxis } from "./formatGraph";
+import { useDeleteFromMonitoring, useOptionsFromMonitoring } from "@/composables/serviceServiceControlUrls";
+import { storeToRefs } from "pinia";
+import { useMonitoringStore } from "@/stores/MonitoringStore";
+import NoData from "@/components/NoData.vue";
+import SmallGraph from "./SmallGraph.vue";
+
+const isRemovingEndpointEnabled = ref(false);
+const router = useRouter();
+const monitoringStore = useMonitoringStore();
+const { historyPeriod } = storeToRefs(monitoringStore);
+const endpoint = defineModel({});
+
+async function removeEndpoint(endpointName, instance) {
+  try {
+    await useDeleteFromMonitoring("monitored-instance/" + endpointName + "/" + instance.id);
+    endpoint.value.instances.splice(endpoint.value.instances.indexOf(instance), 1);
+    if (endpoint.value.instances.length === 0) {
+      router.push({ name: "monitoring", query: { historyPeriod: historyPeriod.value.pVal } });
+    }
+  } catch (err) {
+    console.log(err);
+    return false;
+  }
+}
+
+function navigateToMessageGroup($event, groupId) {
+  if ($event.target.localName !== "button") {
+    router.push({ name: "message-groups", params: { groupId: groupId } });
+  }
+}
+
+async function getIsRemovingEndpointEnabled() {
+  try {
+    const response = await useOptionsFromMonitoring();
+    if (response) {
+      const headers = response.headers;
+      const allow = headers.get("Allow");
+      const deleteAllowed = allow.indexOf("DELETE") >= 0;
+      return deleteAllowed;
+    }
+  } catch (err) {
+    console.log(err);
+  }
+  return false;
+}
+
+onMounted(async () => {
+  isRemovingEndpointEnabled.value = await getIsRemovingEndpointEnabled();
+});
+</script>
+
+<template>
+  <div class="row">
+    <div class="col-xs-12 no-side-padding">
+      <!-- Breakdown by instance-->
+      <!--headers-->
+      <div v-if="loadedSuccessfully" class="row box box-no-click table-head-row">
+        <div class="col-xs-4 col-xl-8">
+          <div class="row box-header">
+            <div class="col-xs-12">Instance Name</div>
+          </div>
+        </div>
+        <div class="col-xs-2 col-xl-1 no-side-padding">
+          <div class="row box-header">
+            <div class="col-xs-12 no-side-padding" v-tooltip :title="`Throughput: The number of messages per second successfully processed by a receiving endpoint.`">Throughput <span class="table-header-unit">(msgs/s)</span></div>
+          </div>
+        </div>
+        <div class="col-xs-2 col-xl-1 no-side-padding">
+          <div class="row box-header">
+            <div class="col-xs-12 no-side-padding" v-tooltip :title="`Scheduled retries: The number of messages per second scheduled for retries (immediate or delayed).`">Scheduled retries <span class="table-header-unit">(msgs/s)</span></div>
+          </div>
+        </div>
+        <div class="col-xs-2 col-xl-1 no-side-padding">
+          <div class="row box-header">
+            <div class="col-xs-12 no-side-padding" v-tooltip :title="`Processing time: The time taken for a receiving endpoint to successfully process a message.`">Processing Time <span class="table-header-unit">(t)</span></div>
+          </div>
+        </div>
+        <div class="col-xs-2 col-xl-1 no-side-padding">
+          <div class="row box-header">
+            <div class="col-xs-12 no-side-padding" v-tooltip :title="`Critical time: The elapsed time from when a message was sent, until it was successfully processed by a receiving endpoint.`">
+              Critical Time <span class="table-header-unit">(t)</span>
+            </div>
+          </div>
+        </div>
+      </div>
+
+      <NoData v-if="endpoint.instances.length == 0" title="No messages" message="No messages processed in this period of time"></NoData>
+
+      <div class="row endpoint-instances">
+        <div class="col-xs-12 no-side-padding">
+          <div class="row box endpoint-row" v-for="(instance, id) in endpoint.instances" :key="id">
+            <div class="col-xs-12 no-side-padding">
+              <div class="row">
+                <div class="col-xs-4 col-xl-8 endpoint-name">
+                  <div class="row box-header">
+                    <div class="no-side-padding lead righ-side-ellipsis" v-tooltip :title="instance.name">
+                      {{ instance.name }}
+                    </div>
+                    <div class="col-lg-4 no-side-padding endpoint-status">
+                      <span class="warning" v-if="formatGraphDuration(instance.metrics.criticalTime).value < 0">
+                        <i class="fa pa-warning" v-tooltip :title="`Warning: instance currently has negative critical time, possibly because of a clock drift.`"></i>
+                      </span>
+                      <span class="warning" v-if="instance.isScMonitoringDisconnected">
+                        <i class="fa pa-monitoring-lost endpoint-details" v-tooltip :title="`Unable to connect to monitoring server`"></i>
+                      </span>
+                      <span class="warning" v-if="instance.isStale">
+                        <i class="fa pa-endpoint-lost endpoint-details" v-tooltip :title="`Unable to connect to instance`"></i>
+                      </span>
+                      <span class="warning" v-if="instance.errorCount" v-tooltip :title="instance.errorCount + ` failed messages associated with this endpoint. Click to see list.`">
+                        <a v-if="instance.errorCount" class="warning cursorpointer" @click="navigateToMessageGroup($event, instance.serviceControlId)">
+                          <i class="fa fa-envelope"></i>
+                          <span class="badge badge-important cursorpointer"> {{ instance.errorCount }}</span>
+                        </a>
+                      </span>
+                    </div>
+                  </div>
+                </div>
+                <div class="col-xs-2 col-xl-1 no-side-padding">
+                  <div class="row box-header">
+                    <div class="no-side-padding">
+                      <SmallGraph :type="'throughput'" :isdurationgraph="false" :plotdata="instance.metrics.throughput" :minimumyaxis="smallGraphsMinimumYAxis.throughput" :metricsuffix="'MSGS/S'" />
+                      <span class="no-side-padding sparkline-value">
+                        {{ instance.isStale == true || instance.isScMonitoringDisconnected == true ? "" : formatGraphDecimal(instance.metrics.throughput) }}
+                        <strong v-if="instance.isStale && !instance.isScMonitoringDisconnected" v-tooltip :title="`No metrics received or instance is not configured to send metrics`">?</strong>
+                        <strong v-if="instance.isScMonitoringDisconnected" v-tooltip :title="`Unable to connect to monitoring server`">?</strong>
+                      </span>
+                    </div>
+                  </div>
+                </div>
+                <div class="col-xs-2 col-xl-1 no-side-padding">
+                  <div class="row box-header">
+                    <div class="no-side-padding">
+                      <SmallGraph :type="'retries'" :isdurationgraph="false" :plotdata="instance.metrics.retries" :minimumyaxis="smallGraphsMinimumYAxis.retries" :metricsuffix="'MSGS/S'" />
+                      <span class="no-side-padding sparkline-value">
+                        {{ instance.isStale == true || instance.isScMonitoringDisconnected == true ? "" : formatGraphDecimal(instance.metrics.retries) }}
+                        <strong v-if="instance.isStale && !instance.isScMonitoringDisconnected" v-tooltip :title="`No metrics received or instance is not configured to send metrics`">?</strong>
+                        <strong v-if="instance.isScMonitoringDisconnected" v-tooltip :title="`Unable to connect to monitoring server`">?</strong>
+                      </span>
+                    </div>
+                  </div>
+                </div>
+                <div class="col-xs-2 col-xl-1 no-side-padding">
+                  <div class="row box-header">
+                    <div class="no-side-padding">
+                      <SmallGraph :type="'processing-time'" :isdurationgraph="true" :plotdata="instance.metrics.processingTime" :minimumyaxis="smallGraphsMinimumYAxis.processingTime" />
+                      <span class="no-side-padding sparkline-value">
+                        {{ instance.isStale == true || instance.isScMonitoringDisconnected == true ? "" : formatGraphDuration(instance.metrics.processingTime).value }}
+                        <strong v-if="instance.isStale && !instance.isScMonitoringDisconnected" v-tooltip :title="`No metrics received or instance is not configured to send metrics`">?</strong>
+                        <strong v-if="instance.isScMonitoringDisconnected" v-tooltip :title="`Unable to connect to monitoring server`">?</strong>
+                        <span v-if="!instance.isStale && !instance.isScMonitoringDisconnected" class="unit">
+                          {{ formatGraphDuration(instance.metrics.processingTime).unit }}
+                        </span>
+                      </span>
+                    </div>
+                  </div>
+                </div>
+                <div class="col-xs-2 col-xl-1 no-side-padding">
+                  <div class="row box-header">
+                    <div class="no-side-padding">
+                      <SmallGraph :type="'critical-time'" :isdurationgraph="true" :plotdata="instance.metrics.criticalTime" :minimumyaxis="smallGraphsMinimumYAxis.criticalTime" />
+                      <span class="no-side-padding sparkline-value" :class="{ negative: formatGraphDuration(instance.metrics.criticalTime).value < 0 }">
+                        {{ instance.isStale == true || instance.isScMonitoringDisconnected == true ? "" : formatGraphDuration(instance.metrics.criticalTime).value }}
+                        <strong v-if="instance.isStale && !instance.isScMonitoringDisconnected" v-tooltip :title="`No metrics received or instance is not configured to send metrics`">?</strong>
+                        <strong v-if="instance.isScMonitoringDisconnected" v-tooltip :title="`Unable to connect to monitoring server`">?</strong>
+                        <span v-if="!instance.isStale && !instance.isScMonitoringDisconnected" class="unit">
+                          {{ formatGraphDuration(instance.metrics.criticalTime).unit }}
+                        </span>
+                      </span>
+                    </div>
+                  </div>
+                </div>
+
+                <!--remove endpoint-->
+                <div class="col-xs-2 col-xl-1 no-side-padding">
+                  <a v-if="isRemovingEndpointEnabled && instance.isStale" class="remove-endpoint" @click="removeEndpoint(endpointName, instance)">
+                    <i class="fa fa-trash" v-tooltip :title="`Remove endpoint`"></i>
+                  </a>
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
+</template>
+
+<style></style>
